requirements.sdxl.txt
fastapi
uvicorn
torch>=2.0.0
numpy
Pillow
diffusers>=0.27.0
accelerate
safetensors
transformers
